create base of error dialog in utils
